+++ game/src/net/protocol/scenario_settings.cpp
@@ -18,7 +18,7 @@ void NetPkg::set_scn_vars(const ScenarioSettings &scn) {
 	if (scn.restricted      ) flags |= 1 << 6;
 
 	write("6I4IB", pkgargs({
-		scn.width, scn.height, scn.popcap, scn.age, scn.seed, scn.villagers,
+		(uint64_t)scn.width, (uint64_t)scn.height, (uint64_t)scn.popcap, (uint64_t)scn.age, (uint64_t)scn.seed, (uint64_t)scn.villagers,
 		(uint64_t)scn.res.food, (uint64_t)scn.res.wood, (uint64_t)scn.res.gold, (uint64_t)scn.res.stone,
 		(uint64_t)flags,
 	}), false);
